# generated/openapi/storyos-public-release-1.yaml
openapi: 3.1.0
info:
  title: StoryOS Public Release 1
  version: storyos.public.release.1
  x-storyos-contract-revision: release1-wire-catalog-2026-09-26-refused-edit-draft
  x-storyos-implemented-slice: getProtocolProfile,getProject,getChapter,createProjectChallenge,createProject,listProjects,updateProject,getProjectAssistance,updateProjectAssistance,createAgentRun,pauseAgentRun,cancelAgentRun,getAgentRun,getRefusedEditDraft,getProposal,acceptProposal,rejectProposalOperations,reopenRejectedOperations,completeReadyPartialProposal,continueProposalGeneration,archiveProject,createVolume,updateVolume,deleteVolume,createChapter,updateChapter,deleteChapter,setCurrentChapter,createProjectCommandChallenge,createEditorSession,getEditorSession,applyAuthorEdit,getApplyAuthorEditOutcome,getSnapshot,getManuscriptTree,searchManuscript,getStatistics,exportHumanReadableManuscript,getHumanReadableManuscriptExport,exportProjectArchive,getExportOperation,activityStream,takeOverProjectWriter,undoLatestAuthorAction
paths:
...
  /api/v1/projects/{project_id}/manuscript/author-edits:
    post:
      operationId: applyAuthorEdit
      summary: Settle one direct Author Edit
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Origin
          in: header
          required: true
          schema:
            type: string
            format: uri
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: X-StoryOS-Anti-Forgery
          in: header
          required: true
          schema:
            type: string
            pattern: '^[0-9a-f]{64}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../json-schema/storyos-public-release-1/apply-author-edit-request.schema.json'
      responses:
        '200':
          description: Committed author edit
          content:
            application/json:
              schema:
                $ref: '../json-schema/storyos-public-release-1/apply-author-edit-response.schema.json'
        '400':
          description: Invalid request
        '401':
          description: Authentication required
        '403':
          description: Request origin refused
        '404':
          description: Resource unavailable
        '405':
          description: Method not allowed
        '409':
          description: Idempotency or authoritative Head conflict
        '412':
          description: Session or writer binding refused
        '413':
          description: Request too large
        '415':
          description: Unsupported content type
        '422':
          description: Author edit refused
        '428':
          description: Precondition required
        '429':
          description: Rate limited
        '503':
          description: Service unavailable
...